refactor: filterFn file name
